PRO-3477 Use prawn-gmagick built from bugfix branch, slim down image
There's an upstream bug in prawn-gmagick, but the maintainer of that project
isn't very active, so it may be a while before it's merged. In the meantime,
the doctools image will install a manually built gem from the branch
containing the fix.

This commit also simplifies the image, removing the jekyll bits (because
they're in advancedtelematic/jekyll-asciidoc) and using a musl-built pandoc so
we don't have to use glibc.